Bank and Basic

+++ axionpro.application/DTOS/Pagination/BaseRequest.cs
@@ -15,9 +15,10 @@ namespace axionpro.application.DTOS.Pagination
         public string? SortOrder { get; set; } = "desc";
         public string? UserEmployeeId { get; set; } 
         public long _UserEmployeeId { get; set; } 
+      
         public string Id { get; set; } = string.Empty;
         public long Id_long { get; set; }
-        public long Id_int { get; set; }
+        public int Id_int { get; set; }
 
     }
